audit_ease/requirements.txt
amqp==5.3.1
asgiref==3.11.0
attrs==25.4.0
billiard==4.2.4
boto3==1.36.5
botocore==1.36.26
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
cssselect2==0.8.0
dj-rest-auth==7.0.2
Django==5.2.10
django-allauth==65.14.0
django-auditlog==3.4.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-ratelimit==4.1.0
django-storages==1.14.2
django-redis==6.0.0
django_celery_results==2.6.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
fonttools==4.61.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jmespath==1.0.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pillow==12.1.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.23
pydyf==0.12.1
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.2
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
s3transfer==0.11.3
sentry-sdk==2.50.0
six==1.17.0
sqlparse==0.5.5
stripe==10.8.0
tinycss2==1.5.1
tinyhtml5==2.0.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
weasyprint==68.0
webencodings==0.5.1
zopfli==0.4.0
